--- a/packages/jupyter-ai-magics/jupyter_ai_magics/embedding_providers.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/embedding_providers.py
@@ -38,8 +38,6 @@ class BaseEmbeddingsProvider(BaseModel):
 
     model_id: str
 
-    provider_klass: ClassVar[Type[Embeddings]]
-
     registry: ClassVar[bool] = False
     """Whether this provider is a registry provider."""
 
@@ -47,28 +45,38 @@ class BaseEmbeddingsProvider(BaseModel):
     """Fields expected by this provider in its constructor. Each `Field` `f`
     should be passed as a keyword argument, keyed by `f.key`."""
 
+    def __init__(self, *args, **kwargs):
+        try:
+            assert kwargs["model_id"]
+        except:
+            raise AssertionError(
+                "model_id was not specified. Please specify it as a keyword argument."
+            )
 
-class OpenAIEmbeddingsProvider(BaseEmbeddingsProvider):
+        model_kwargs = {}
+        model_kwargs[self.__class__.model_id_key] = kwargs["model_id"]
+
+        super().__init__(*args, **kwargs, **model_kwargs)
+
+class OpenAIEmbeddingsProvider(BaseEmbeddingsProvider, OpenAIEmbeddings):
     id = "openai"
     name = "OpenAI"
     models = ["text-embedding-ada-002"]
     model_id_key = "model"
     pypi_package_deps = ["openai"]
     auth_strategy = EnvAuthStrategy(name="OPENAI_API_KEY")
-    provider_klass = OpenAIEmbeddings
 
 
-class CohereEmbeddingsProvider(BaseEmbeddingsProvider):
+class CohereEmbeddingsProvider(BaseEmbeddingsProvider, CohereEmbeddings):
     id = "cohere"
     name = "Cohere"
     models = ["large", "multilingual-22-12", "small"]
     model_id_key = "model"
     pypi_package_deps = ["cohere"]
     auth_strategy = EnvAuthStrategy(name="COHERE_API_KEY")
-    provider_klass = CohereEmbeddings
 
 
-class HfHubEmbeddingsProvider(BaseEmbeddingsProvider):
+class HfHubEmbeddingsProvider(BaseEmbeddingsProvider, HuggingFaceHubEmbeddings):
     id = "huggingface_hub"
     name = "HuggingFace Hub"
     models = ["*"]
@@ -78,5 +86,4 @@ class HfHubEmbeddingsProvider(BaseEmbeddingsProvider):
     # tqdm is a dependency of huggingface_hub
     pypi_package_deps = ["huggingface_hub", "ipywidgets"]
     auth_strategy = EnvAuthStrategy(name="HUGGINGFACEHUB_API_TOKEN")
-    provider_klass = HuggingFaceHubEmbeddings
     registry = True