--- a/packages/jupyter-ai-magics/jupyter_ai_magics/embedding_providers.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/embedding_providers.py
@@ -58,6 +58,7 @@ class BaseEmbeddingsProvider(BaseModel):
 
         super().__init__(*args, **kwargs, **model_kwargs)
 
+
 class OpenAIEmbeddingsProvider(BaseEmbeddingsProvider, OpenAIEmbeddings):
     id = "openai"
     name = "OpenAI"