--- a/packages/jupyter-ai-magics/jupyter_ai_magics/magics.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/magics.py
@@ -11,7 +11,7 @@ import click
 from IPython import get_ipython
 from IPython.core.magic import Magics, line_cell_magic, magics_class
 from IPython.display import HTML, JSON, Markdown, Math
-from jupyter_ai_magics.utils import decompose_model_id, load_providers
+from jupyter_ai_magics.utils import decompose_model_id, get_lm_providers
 from langchain.chains import LLMChain
 
 from .parsers import (
@@ -138,7 +138,7 @@ class AiMagics(Magics):
             "`from langchain.chat_models import ChatOpenAI`",
         )
 
-        self.providers = load_providers()
+        self.providers = get_lm_providers()
 
         # initialize a registry of custom model/chain names
         self.custom_model_registry = MODEL_ID_ALIASES