--- a/packages/jupyter-ai-magics/jupyter_ai_magics/providers.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/providers.py
@@ -17,7 +17,6 @@ from langchain.llms import (
 )
 from langchain.llms.sagemaker_endpoint import LLMContentHandler
 from langchain.llms.utils import enforce_stop_tokens
-from langchain.schema import BaseModel as BaseLangchainProvider
 from langchain.utils import get_from_dict_or_env
 from pydantic import BaseModel, Extra, root_validator
 
@@ -66,7 +65,7 @@ class MultilineTextField(BaseModel):
 Field = Union[TextField, MultilineTextField]
 
 
-class BaseProvider(BaseLangchainProvider):
+class BaseProvider(BaseModel):
     #
     # pydantic config
     #