--- a/packages/jupyter-ai-magics/jupyter_ai_magics/utils.py
+++ b/packages/jupyter-ai-magics/jupyter_ai_magics/utils.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Dict, Optional, Tuple, Union
+from typing import Dict, Optional, Tuple, Union, Type
 
 from importlib_metadata import entry_points
 from jupyter_ai_magics.aliases import MODEL_ID_ALIASES
@@ -7,9 +7,11 @@ from jupyter_ai_magics.embedding_providers import BaseEmbeddingsProvider
 from jupyter_ai_magics.providers import BaseProvider
 
 Logger = Union[logging.Logger, logging.LoggerAdapter]
+LmProvidersDict = Dict[str, BaseProvider]
+EmProvidersDict = Dict[str, BaseEmbeddingsProvider]
+ProviderDict = Dict[str, Union[BaseProvider, BaseEmbeddingsProvider]]
 
-
-def load_providers(log: Optional[Logger] = None) -> Dict[str, BaseProvider]:
+def get_lm_providers(log: Optional[Logger] = None) -> LmProvidersDict:
     if not log:
         log = logging.getLogger()
         log.addHandler(logging.NullHandler())
@@ -30,10 +32,9 @@ def load_providers(log: Optional[Logger] = None) -> Dict[str, BaseProvider]:
 
     return providers
 
-
-def load_embedding_providers(
+def get_em_providers(
     log: Optional[Logger] = None,
-) -> Dict[str, BaseEmbeddingsProvider]:
+) -> EmProvidersDict:
     if not log:
         log = logging.getLogger()
         log.addHandler(logging.NullHandler())
@@ -75,3 +76,18 @@ def decompose_model_id(
 
     provider_id, local_model_id = model_id.split(":", 1)
     return (provider_id, local_model_id)
+
+def get_lm_provider(model_id: str, lm_providers: LmProvidersDict) -> Tuple[str, Type[BaseProvider]]:
+    """Gets a two-tuple (<local-model-id>, <provider-class>) specified by a
+    global model ID."""
+    return _get_provider(model_id, lm_providers)
+
+def get_em_provider(model_id: str, em_providers: EmProvidersDict) -> Tuple[str, Type[BaseEmbeddingsProvider]]:
+    """Gets a two-tuple (<local-model-id>, <provider-class>) specified by a
+    global model ID."""
+    return _get_provider(model_id, em_providers)
+
+def _get_provider(model_id: str, providers: ProviderDict):
+    provider_id, local_model_id = decompose_model_id(model_id, providers)
+    provider = providers.get(provider_id, None)
+    return local_model_id, provider