--- a/packages/jupyter-ai/jupyter_ai/actors/__init__.py
+++ b/packages/jupyter-ai/jupyter_ai/actors/__init__.py
@@ -1,1 +1,0 @@
-"""Actor classes that process incoming chat messages"""