--- a/packages/jupyter-ai/jupyter_ai/chat_handlers/__init__.py
+++ b/packages/jupyter-ai/jupyter_ai/chat_handlers/__init__.py
@@ -1,0 +1,6 @@
+from .ask import AskChatHandler
+from .base import BaseChatHandler
+from .clear import ClearChatHandler
+from .default import DefaultChatHandler
+from .generate import GenerateChatHandler
+from .learn import LearnChatHandler