… .github/workflows/check-release.yml
… .github/workflows/lint.yml
… .github/workflows/prep-release.yml
… .github/workflows/publish-release.yml
… README.md
… docs/source/contributors/index.md
… docs/source/users/index.md
… examples/generate/Creating Random Arrays with Numpy.ipynb
… examples/generate/Introduction to Sets.ipynb
… examples/generate/Python Sets Tutorial.ipynb
… examples/generate/Training Dense Neural Network with PyTorch.ipynb
… examples/generate/Training a Dense Neural Network with PyTorch.ipynb
… packages/jupyter-ai-magics/pyproject.toml
… packages/jupyter-ai/README.md
… packages/jupyter-ai/jupyter_ai/actors/memory.py
… packages/jupyter-ai/jupyter_ai/actors/providers.py
… packages/jupyter-ai/jupyter_ai/actors/router.py
… packages/jupyter-ai/jupyter_ai/chat_handlers/ask.py
… packages/jupyter-ai/jupyter_ai/chat_handlers/base.py
… packages/jupyter-ai/jupyter_ai/chat_handlers/clear.py
… packages/jupyter-ai/jupyter_ai/chat_handlers/default.py
… packages/jupyter-ai/jupyter_ai/chat_handlers/generate.py
… packages/jupyter-ai/jupyter_ai/chat_handlers/learn.py
… packages/jupyter-ai/jupyter_ai/config_manager.py
… packages/jupyter-ai/jupyter_ai/document_loaders/directory.py
… packages/jupyter-ai/jupyter_ai/engine.py
… packages/jupyter-ai/jupyter_ai/extension.py
… packages/jupyter-ai/jupyter_ai/handlers.py
… packages/jupyter-ai/jupyter_ai/models.py
… packages/jupyter-ai/jupyter_ai/reply_processor.py