# Description

Replaces Ray with asyncio where possible and Dask where necessary.

Additional changes:
- Raises Python version ceiling to 3.11
- Bumps LangChain dependency to v0.0.220
- Updates documentation
- Removes more unused code from the era of model engines
  - Task API, Prompt API, TaskManager, Inserters.

Issues closed:

- Closes #242
- Closes #235

# Demo

This demonstrates how multiprocessing is not necessary to implement concurrency and non-blocking messages.

https://github.com/jupyterlab/jupyter-ai/assets/44106031/ec6be1d7-bcc8-47c2-a7e4-b8da59a7f895

# Performance profiling

All performance profiling was done on a M1 Macbook Pro with 32 GB unified memory. Each task was performed with the following prompts:

| Task                        | Prompt                     |
| --------------------------- | -------------------------- |
| Initialize server extension | N/A                        |
| Default                     | "Hello!"                   |
| /ask                        | "What is Jupyter AI?"      |
| /learn                      | "./docs"                   |
| /generate                   | "a notebook about asyncio" |

`main` was profiled using #246, which adds performance profiling on `main`.

### Task latency on `main`

| Task                        | Run 1  | Run 2 | Run 3 | Avg. Latency (ms) | Avg. Latency (s) |
| --------------------------- | ------ | ----- | ----- | ----------------- | ---------------- |
| Initialize server extension | 2194   | 2211  | 2218  | 2208              | 2.208            |
| Default                     | 2091   | 1129  | 1186  | 1469              | 1.469            |
| /ask                        | 4663   | 14452 | 17830 | 12315             | 12.315           |
| /learn                      | 13459  | 12856 | 10248 | 12188             | 12.188           |
| /generate                   | 269631 | 76938 | 87915 | 144828            | 144.828          |

### Task latency on `dask` (current branch)

| Task                        | Run 1 | Run 2 | Run 3 | Avg. Latency (ms) | Avg. Latency (s) |
| --------------------------- | ----- | ----- | ----- | ----------------- | ---------------- |
| Initialize server extension | 22    | 21    | 21    | 21                | 0.021            |
| Default                     | 2125  | 1117  | 617   | 1286              | 1.286            |
| /ask                        | 4746  | 13394 | 10390 | 9510              | 9.51             |
| /learn                      | 2446  | 907   | 1460  | 1604              | 1.604            |
| /generate                   | 28099 | 55155 | 51436 | 44897             | 44.897           |

### Summary of performance profiling

| Task                        | Avg. Latency (ms, `main`) | Avg. Latency (ms, `dask`) | Speedup Factor |
| --------------------------- | --------------------------- | --------------------------- | -------------- |
| Initialize server extension | 2208                        | 21                          | 105.1          |
| Default                     | 1469                        | 1286                        | 1.1            |
| /ask                        | 12315                       | 9510                        | 1.3            |
| /learn                      | 12188                       | 1604                        | 7.6            |
| /generate                   | 144828                      | 44897                       | 3.2            |

- Server initialization time was greatly decreased by a very large constant term because this branch no longer forks/spawns processes on init.
- `/learn` and `/generate` enjoy a generous speedup thanks to rewriting actors handling network requests serially to chat handlers handling network requests concurrently.
- The default and `/ask` latency were decreased only by an apparently constant amount of time, as the performance gain is due exclusively to removing IPC overhead.

# Developer guidance

### Where did the actors go?

Previously, the actors had a complex call flow:

<img width="1114" alt="Screenshot 2023-06-29 at 11 01 57 AM" src="https://github.com/jupyterlab/jupyter-ai/assets/44106031/2ace04e8-7a9d-4e09-a01d-2e729001a51d">

<details>
<summary>Mermaid code</summary>

```
graph TD;
    RCH["ChatHandler (root chat handler)"];
    CH[GlobalConfigHandler];
    R[RouterActor];
    A[AskActor];
    L[LearnActor];
    G[GenerateActor];
    D[DefaultActor];
    C[ConfigActor];
    M[MemoryActor];
    CP[ChatProviderActor];
    EP[EmbeddingsProviderActor];
    P[ProvidersActor];
    
    RCH --> R;
    R --> A & L & G & D & C;
    CH --> C;
    D --> M;
    A --> CP;
    L --> EP;
    CP & EP --> P;
    C --> CP & EP;
```

</details>

Here is how this PR migrates each actor:
- Router actor is now implemented into the `_route()` method on the root chat handler, i.e. the one defined in `handlers.py`
- Ask, Learn, Generate, Default are now implemented as _chat handlers_ under `chat_handlers/`.
  - Base actor now is `BaseChatHandler`, defined in `chat_handlers/base.py`
- Config actor is now implemented in `ConfigManager`, defined in `config_manager.py`
- Chat provider, Embeddings provider, and Provider actors were all merged into `ConfigManager`
  - This transformation is quite natural; retrieving a provider is similar to "deserializing" a field from the configuration, which implies that this type of operation should be a method on whatever class is wrapping the configuration.
- Memory actor is now merged into `DefaultChatHandler`, which is fine since nobody else was using it anyways.

The call flow now looks like this:

<img width="1116" alt="Screenshot 2023-06-29 at 11 02 44 AM" src="https://github.com/jupyterlab/jupyter-ai/assets/44106031/c4c4c3af-14b2-4ae2-ae0c-81dad5be1acf">

<details>
<summary>Mermaid code</summary>

```
graph TD;
    RCH[RootChatHandler];
    CH[GlobalConfigHandler];
    A[AskChatHandler];
    L[LearnChatHandler];
    G[GenerateChatHandler];
    D[DefaultChatHandler];
    C[ClearChatHandler];
    CM[ConfigManager];

    
    RCH --> A & L & G & D & C;
    CH --> CM;
    A & L & G & D --> CM;
```

</details>

### Why does this work?

This works because it's not necessary to spawn multiple processes to handle web server concurrency. Tornado itself is a single-threaded web server. Async coroutines handle concurrency very well thanks to how fast the context switches are relative to using processes. We should reserve parallelism for CPU-bound/GIL-bound parallelizable tasks (as is the case for document indexing), and prefer `asyncio` everywhere else, as typically we are I/O-bound and just need concurrency.

### How should I review this PR?

I recommend testing all commands and use-cases, making sure they're concurrent (i.e. allow the backend to respond to a quick message after a pending message, e.g. `/generate`), and then making sure that all Python versions are known to work.