@JasonWeill 2023-06-29
By adding Python 3.11 support, this would fix #235.