@dlqqq 2023-06-30
@3coins Thanks for the callout. I traced the bug to https://github.com/hwchase17/langchain/pull/4500, which was fixed in v0.0.172. I've upgraded our LangChain dependency to the latest version, v0.0.220. I tested this locally with Python 3.9, and everything still works.